inverted: move the batched Contains fold to its own file
searcher_doc_bitmap.go holds the docBitmap builders for every index shape — set,
map, roaring set, range, geo — and the batched Contains fold had grown to a third
of it while sharing nothing with the rest. The work that follows adds a planner and
a parallel runner to the fold alone, which would bury those builders.

Code and tests move unchanged; the file split is the whole diff. The only edits are
the two imports the remaining builders no longer use.
